Fix: Cast damage and hp values to int where needed

+++ Source By Mr Blue/src/nro/models/Bot/BotAttackplayer.java
@@ -224,7 +224,7 @@ public class BotAttackplayer extends Bot {
 
         int damePST = (int) (dame * percentPST / 100L);
         if (damePST >= attacker.nPoint.hp) {
-            damePST = attacker.nPoint.hp - 1;
+            damePST = (int) (attacker.nPoint.hp - 1);
         }
 
         Message msg = null;
@@ -232,7 +232,7 @@ public class BotAttackplayer extends Bot {
             msg = new Message(56);
             msg.writer().writeInt((int) attacker.id);
             damePST = attacker.injured(attacker, damePST, true, false);
-            msg.writer().writeInt(attacker.nPoint.hp);
+            msg.writer().writeInt((int) attacker.nPoint.hp);
             msg.writer().writeInt(damePST);
             msg.writer().writeBoolean(false);
             msg.writer().writeByte(36);
